Initial commit of the ailuropoda project
This project is to create an API endpoint for a sad panda. We have a
wiki[1], launchpad[2], blueprints[3] set up already.

Some of the larger enterprise customers requested a visual representation
of an error page. This is a standard API endpoint for all of OpenStack
to use to have the standard error. Initial release is just an image, but
eventually we plan on having a more dynamic way to serve pictures.

This is Sad Pandas as a Service.

[1]: http://wiki.openstack.org/wiki/ailuropoda
[2]: https://launchpad.net/ailuropoda
[3]: https://blueprints.launchpad.net/ailuropoda

Change-Id: I30a6e47bf0edd146d9fb762d66ec0a1e248ec3bf