Get translation jobs into healthcheck
Add subunit testr output at end of run of the translation scripts, so
that healthcheck will have information about success/failure of the job.

This is implemented as an exit trap since we abort on every error and
need to provide the failure message.

Also, upload the healthcheck data to logs.openstack.org so that
healthcheck can import it.

This should be a temporary solution until healthcheck can get data from
Zuul about failure/success of job runs using database access.

Change-Id: I74bba9659a9767027e2daa649ec6f38bf822b65a