Allow tox targets that need sudo (functional jobs)
In python-k8sclient functional job, we need to sudo
perms to setup docker and then run k8s containers in
docker before we can run the tests. We cannot use
dsvm jobs as we are not really setting up openstack.
We are just testing kubernetes client and need to
setup/start kubernetes (as docker containers).

Change-Id: I4ddb6098ba2787a82f134619bcda077da49077a1