Add oslo.tools repo/project
The oslo-incubator is no longer supported and the piece
that is used from it is some of the tooling inside of it.

These tools were extracted at:

https://github.com/harlowja/oslo.tools.git

And this proposes a tiny new project just for these tools
that others in the community can benefit from (and allows the
oslo group to put the incubator in the attic).

Needed-By: I818886696aee39c2fef786f21c59202b4ced1dbe
Change-Id: Id5ed1a47c9bf1e4b25557b8ce802ef01694a2780