Add non-voting docs job for fuel-web
Enable building of Nailgun API docs in fuel-web by upstream infra.
This is a part of initiative to move all documentation related processes
upstream for fuel-* repos and deprecate some Fuel CI jobs in favour of
upstream ones.
Docs building should work fine for fuel-web, but let's enable it in
non-voting to ensure everything is ok.

Change-Id: Idda045fb291b1ab80b8fd6e20d4af7e40b1cc48e