Revert "Changing gate on devstack identity v3 only voting"
This job was set as a voting gate job on only devstack and tempest.
However, there was no equivalent coverage in check (there was only a
nv v3 only job) Additionally, there were no other projects with a
voting v3 only job. This commit removes this as it's clearly not
something we should be gating on and leaves us open to wedging either
tempest or devstack with a change in one of the projects.

This reverts commit aed321fae933b6b59153fef80068343d3b60cbb5.

Change-Id: I0c1e0520ee1a1234d7fbea98e2e04d3c29885610