create os-api-ref project
This adds a new project which is a sphinx extension that assists in
the creation of API Reference Sites for OpenStack projects. This was
POCed in the Nova tree, and is being extracted into an installable
library so that other projects can copy this model without
copy/pasting the Nova extension into their tree.

Change-Id: Id2147db7603dc9c3a5d0e9df72e2ce7e4f4fcbf9