Increase horizon integration tests timeout
Right now we have a problem, that our integration tests are failed
often due to build timeout. It happens because we have 60 tests
which require 30-90 sec per one. First of all we would like to
up build timeout to prevent flaky timeout failures. And then
we are planning to use selenium grid and parallel execution
to provide tests scalability.

Change-Id: I28b08cde1e8c8dff2cde0fa4c12d852373d946a9