tempest: move puppet jobs from exp to check pipeline
We've recently noticed that some patches in Tempest used to break
downstream CIs like Puppet OpenStack, because we deploy in another way
than Devstack:
http://git.openstack.org/cgit/openstack/puppet-openstack-integration/tree/README.md

We used to have Puppet jobs part of experimental pipeline for Tempest
testing but this is not enough.

This is a proposal to run Puppet jobs in the check pipeline, as
non-voting. The jobs will bring valuable feedback in the future without
slowing down Tempest testing (3 jobs, ~30 minutes each).

Change-Id: I46ea5b0cf0e52a0ddc6bac14d1db4a42ae79a1bb