Remove periodic-db-kilo job from kilo-eol projects
This removes the periodic-db-kilo job from the projects
in http://releases.openstack.org/kilo/index.html that
have a kilo-eol tag (note the stable/kilo branch in those
repos have not yet been removed, but will be), with the
exception of murano which was not managed by the stable-maint-team
for releases in kilo. However, this change still removes the
periodic-db-kilo job from openstack/murano since we shouldn't be
supporting kilo for that project anymore either, it's just a
matter of the murano team doing the tag and final release
for kilo.

Change-Id: Iaba84d35ff16e74d21fc91dd74006abafa3511f1