Remove kilo bitrot job definition and usage
Kilo is EOL, this change removes the bitrot job
definitions and usage from the projects using it.

There are a few in here that aren't managed by the
stable-maint-team for the final kilo release, like
manila, murano, zaqar and the neutron-*aas projects,
but we can't speak for the state of their kilo branch
at this point anyway, and they just need to end of life
the branch after tagging it.

Change-Id: I46fb0305bd0187eae2a81828916e9f946f60f1e1