Only publish latest image to magnum
Currently we were generating images daily, and timestamping it,
also creating the latest one with fedora-atomic-f23-dib name.

This was causing that directory was filled with images daily,
growing in size. There is currently no cleanup process available,
so this cannot be maintained.

So instead of that, we will only publish the latest image, after
validation, with a fixed name. Images are validated and not copied
if they are not successful, so this will be safe to land.

Change-Id: I7fcbb6bed034557969067366e1dc2f51e9fbb1b8