translations: Handle not yet translated repos better
Projects that are not translated get an error when running the propose
translation. This is not critical since there's nothing to propose but
let's handle it gracefully so that the exit code is meaningfull.

In this case, we run git add on a path that does not exist, check first
whether it exists.

Change-Id: I08f146b57af1713bf489adbe6f2282ca7d72b259