Add TripleO 2-node job
Adds a new CI job in the experimental-tripleo queue using multinode
nodepool jobs. The job uses 1 subnode to deploy via TripleO. The
jenkins slave is purposed as the undercloud, and an all in one Overcloud
is deployed on the subnode.

Note that support for multinode jobs is being added to tripleo-ci in this
patch: Icc67a289df207348536c0a2419b3fc81f03fe758

Change-Id: I8375a96a4f4394b0a0d233731f7f4b386c0d5c71