disable voting on puppet-ceph beaker jobs
Puppet OpenStack group is working on upgrading puppet-ceph to support
Ubuntu Xenial & Ceph Jewel.
We're using Puppet OpenStack Integration, Fuel and TripleO jobs to
validate it works fine. Though we have some issues with beaker jobs that
we want to solve later. Until this time, let's disable voting on them
now, but keep them visible so we can iterate on it.

Change-Id: I2a09eeccfc92bff926998c1fb2a1d658371bc46f