New Project Quark Plugin
The goal of this is to provide an alternative to ML2 that
is scalable beyond the limits of ML2's capabilities today.
This is an unofficial plugin today but could be added under
stadium in the future.  But, to do so based on current
requirements, the advancement of quark needs to be done
out in the open, community based before it can be considered.

Change-Id: Id4f6aa2a2585750c9176f93bfa8927e5796ba6d3