Don't clear-out all repos in puppet cleanup
dib 1.17.0 includes fixes for debian-minimal to work with the
apt-sources element in I69dbaa34be3db3d667e6bd8450ef4ce04a751c70

This moved to having the base repos split out from the usual
/etc/apt/sources.list into separate files in /etc/apt/sources.list.d

Unfortunately, this puppet module does the cleanup by glob removal of
everything in that directory, which is a bit unsafe.

I've refactored this slightly so that, like the RedHat path, we only
remove the puppetlabs repos files.

Change-Id: I5bcd8880a90d238b77aaacfd1eaf0a720552c7ee