Add a build for the gerrit its-storyboard plugin
The its-storyboard plugin will allows us to integrate gerrit with
storyboard.  Add a build for the gerrit its-storyboard plugin[1].
Building this plugin also requires building the its-base plugin[2]
because the its-storyboard depends on its-base.  Refactored the build
job to build a parent project if it's defined in the project yaml.

[1] https://gerrit.googlesource.com/plugins/its-storyboard
[2] https://gerrit.googlesource.com/plugins/its-base

Change-Id: I7b8461bb47ba902158531a64ddf88a9bb2b03312