Update OSA job skip for all jobs except docs/releasenotes
This updates the job skip for OpenStack-Ansible repositories to
ensure that all jobs except docs/releasenotes are skipped if the
content in the patch only changes docs/releasenotes.

This change specifically changes this skip to act against all
OSA repositories (except the -specs repository) rather than just
the openstack/openstack-ansible repository.

For testing, the following jobs should match, and then be skipped if
the patch content only contains the designated docs changes:
- gate-openstack-ansible-dsvm-commit
- gate-openstack-ansible-linters
- gate-openstack-ansible-rabbitmq_server-linters
- gate-openstack-ansible-rabbitmq_server-ansible-func-centos-7-nv
- gate-openstack-ansible-rabbitmq_server-ansible-func-ubuntu-trusty
- gate-openstack-ansible-rabbitmq_server-ansible-func-ubuntu-xenial-nv

And the following jobs should not match, and therefore always execute:
- gate-openstack-ansible-docs
- gate-openstack-ansible-releasenotes
- gate-openstack-ansible-rabbitmq_server-docs
- gate-openstack-ansible-rabbitmq_server-releasenotes

Change-Id: Iddd56c6eef01e3e114becf0d09206d734b6c3c55