Merge gerrit jobs for project-config
Both gate-project-config-projects-yaml and gate-project-config-gerrit
test files in gerrit directory, both check ACLs a swell. Let's merge
those into one job since in most cases both need to be run anyway and
the name projects-yaml might be confused with
jenkins/jobs/projects.yaml.

Note that this one leaves the projects tox environment in so that
testing of this change passes.

Change-Id: I34324643c26d0bcf2d24ce1dc51250ad53f2d48f