Fix preseed script for the brave new world
Ever since we started populating the subunit2sql db with data from all
gate and periodic jobs this script has been kinda useless. This was
added with the intent to preseed tempest runs with data from past runs
to hopefully have a positive impact on the testr scheduler. (its
actual usefulness is limited because of bug 1416512 and other
scheduler limitations) However, because we're not filtering now we
end up preseeding tempest runs with data from things like zuul unit
tests which are just a waste. This commit adds the necessary
filtering to the DB queries to ensure we only preseed with data
from previous tempest runs.

Note: For this patch to work the depends-on below must be included in
a subunit2sql release

Change-Id: I32866d63b5f0fa406213709fc88a9029d0318645
Depends-On: Id6ca7e7a4cd1f913d9f84efae5f48d247dd9ed64