Add trove-dashboard to project list on layer4 job
The trove devstack plugin recently added trove-dashboard rep to the
list of things installed but neglected to update the tempest job which
runs trove. This has broken all tempest dev because we can't clone the
trove-dashboard repo. This commit adds the missing the repo to the
PROJECTS list and should unblock things.

Change-Id: I816574051d48f8d75894a8cdf010294080a794eb