Add repositories for Charm interfaces
Create repositories and perform initial import of a number
of Charm interface modules associated with the OpenStack
Charm set; this change will be followed by a set of layers
and then charms for Aodh, Designate and Barbican which
consume all of these resources.

Interfaces are relatively simple right now and only include
pep8 checks; Unit tests are due to be added soon.

Change-Id: I55f1c071d739955c7604f4af5bef306992268c4a