Fix gerrit its-storyboard build
The build was failing because it was attempting to clone its-base
plugin with the same branch as the gerrit core.  This will typically
work for upstream builds but since we have a fork of gerrit core
(our own branch[1]) it doesn't work.  This change lets us configure
which branch to clone from, in this case just clone from the same
branch[2] as the plugin.

[1] http://git.openstack.org/cgit/openstack-infra/gerrit/log/?h=openstack/2.11.4
[2] https://gerrit.googlesource.com/plugins/its-base/+/stable-2.11

Change-Id: Ie20ef8f6e9f4500772bcc16f305c9c9e6466c44e