Enable file injection in neutron full ssh job
This was previously enabled in tempest but was broken
by default since the libvirt driver disables file
injection by default, so this job fails because it
ssh's into the guest to verify the injected files exist,
which they wouldn't by default.

Nova is being fixed to fail a server create request if
it can't inject files, and devstack is being updated to
configure the stack to test this properly.

Related-Bug: #1598581

Change-Id: I51ed3bca4e27e82d75c063a7c2a6412174216d14
Depends-On: I34790fadeffd6e3fdc65bd9feed3d6e62316896c