Add job configuration for unit testing against neutron-lib source
Testing neutron-lib periodically with unit test jobs from  Neutron stadium
projects can help prevent breakages detected at the time of a neutron-lib
release and it will speed development and maturity of the library.

This patch introduces the configuration necessary for running unit
tests against the master version of neutron-lib. The support is limited
to py34, as that would suffice to ensure coverage.

Change-Id: Ib1e4dbf92aab4ec09ad3ce54255ca729941784fa