[zmq] Use proper driver name to configure devstack
Devstack was not configured properly (redis-server was not installed)
because driver name was specified with suffix (-proxy and -pub-sub).

In this patch use driver name 'zeromq' for each job, and ostype
suffix was used for suffixes, since job-suffix (which is probably most suitable choice
for this task) is not used for the current template.

The alternative way is to create one more job template for oslo functional tests
which uses job-suffix and apply the template for these particular jobs.

We can also modify current f-tests template to make it use job-suffix,
but then we need to update all oslo jobs defenitions over zuul layout.

Change-Id: I3756624dcb51895ba972e7e9b3692e41ace798f3