Refactor Ironic job macro
Ironic is getting more and more jobs to test the different deployment
scenarios we support, every time we add a new optional feature we add a
new conditional to the existing macro to enable it for one job, and then
we have to add a new variable to every job template that uses that
macro.

This is getting unmanageable and messy, so this patch refactors out the
configurations we've been controlling using conditionals into their own
builders which can be pulled into a job-template only if required. It
does this by writing the optional script parts into a file call
ironic-extra-vars which is then sourced in the main ironic macro.

This patch takes all the Ironic jobs that are largely the same, and
refactors them into a single template, and then uses the jenkins job
builder matrixing system to generate the jobs instead of explicitly
defining each one.

Change-Id: Ie89fb4d481d8c88f433a25587d6447bf4e9e1434