Fixed target path
Paul wrote a Debian repository publisher post job, but it's using
deb-openstack-pkg-tools instead of openstack-pkg-tools as the
path from where to pick-up the package. Therefore, the files should
be pushed to /packaging-deb/deb-openstack-pkg-tools (ie: with deb-
in front) for it to work. Hopefully, this will be the last fix
of upload path.

Change-Id: I658ce15d737e10658fcbddba0188d83b838994c1