Revert "Run identity v3 only job in nova"
This reverts commit b8d80d502d1ada46d231c6edf0578d825ffe8455.

There was no sign off from the Nova PTL or anyone from the
Nova core team about running a non-voting keystone v3-only
job on all nova changes.

If keystone v2 is deprecated, then devstack should be
changed to run keystone v3 only by default (which I
actually think was tried around the Newton summit but
things started failing pretty hard).

Anyway, if this was just a one-time test, then it looks
like it's passing and things are fine.

If someone wants this for further verification while a
long-term plan is being worked out for the keystone API
support story, then this job should be in the experimental
queue, not the check queue.

Change-Id: If7d17ca2e5bfef9b47ccab68e8c4195a39ca8e86