Add new project RSC to openstack repo
RSC is a controller service for lifecycle management of pooled bare-metal
hardware infrastructure (like Rack Scale Design [1]) which uses Redfish as
one of the management protocols.

This project proposes to create a management service utilizing Redfish APIs
in a form that can be used by Ironic, Nova, Neutron or any other
service/controller within Openstack.

RSC is not an official OpenStack project yet.

[1] http://www.intel.com/content/www/us/en/architecture-and-technology/intel-rack-scale-architecture.html

Change-Id: I8010ff9d72393c2c7759c563fddebe2fde3daeb9