Add userstory-dashboard repos
Add the user stories dashboard PoC as an Infra project, and also a
new Puppet module for deploying the same. The Product Working Group
intends to use this as an interim means of tracking progress on user
stories across multiple projects while they help implement the
"epics" model in Storyboard to provide similar features.

Needed-By: I5f2b7d863f208b393cdaa90d2b540a3a7727f197
Change-Id: Ic818dd5c754fe3ed0166942187b91b9d425f4274