Reduce the OSIC max-servers
The OSIC cloud1 environment is currently running out of IPv4 space.
We are working on getting more IPv4 addresses and potentially
adding IPv6 however that will take a bit more time. Reducing this
value will ensure we're not seeing so many errors in the OSIC
as nodepool allocates resources for the gate.

Change-Id: I9c5eb20d965010c140880b1b309871fad57d8782
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>