Renamed *-npm-* jobs to *-nodejs4-npm-* jobs
In order to provide a migration path for projects to move from node4
to node6 and later, this patch renames all npm-based jobs to explicitly
describe which version of node they're running on. The npm-upload builder
was not changed, as this one executes with --no-scripts and therefore
should never depend on the running version of node. Also, it runs on the
release slave, so we cannot dynamically change the nodejs version.

Change-Id: I7eb9cddd689646b8eb8eeee69f0fd5c82d827754