Remove the check tox env from irc-meetings
Initially we added 'check' as a second venv to run during the ical job.
The motivation here was to avoid using a second node to keep the impact low.

This was problematic as while the check was fine, some of the tools assume
that 'envlist' was in fact a single venv.

Also recently we started running pep8 as a second job with an additional
node.

With that in mind remove the check tox env from the job definition and
in irc-meetings we'll treat pep8 as a more generic linter environment

Change-Id: I7ddfa03606aa750308e2e8c6b9c77e9f226bead1