Add diskimage-builder nodepool test
Add a add a "nodepool with diskimage-builder from source" job as
experiemental (and non-voting) for dib.

The idea is to test that new diskimage-builder changes still work with
nodepool to make a bootable image.

(ps. changeset 1 made the shade-src template less specific, but
PROJECTS is different; shade in openstack-infra/shade, dib in
openstack/diskimage-builder.  So the template started looking odd with
too many arguments).

Change-Id: I28356d346117cd571c35a1374aab7bc9031db8c3
Depends-On: I480dcb201aa93bc3d208e2bf5e96bcee6dd7df09