Add automatic backports
To be able to build OpenStack in Debian Stable, we need a bunch of
backported packages from Sid, which aren't maintained under the
Debian PKG OpenStack group.

This patch adds the new deb-auto-backport Git repository which
contains the list of such packages, and the necessary scripting
to build them and push them into an "uploads" folder, which content
is later on (in a post job) uploaded to the Debian repository.

Change-Id: I1e756ae77a9e4b44f2a57201aa162f7e2f51e605