Move gate-{name}-chef-rake to ubuntu-xenial
This commit alters the chef jobs in jenkins to run on xenial in addition
to trusty. In the future the trusty jobs can be removed. Older branches
will continue to use trusty.

This job also renames some chef jobs to have the {node} portion at the
end of the job name, which should fit the job name standard closer.

Change-Id: Idfbb264f31e218d4b3c6f2351175fff5f0d52f8a