Created DSVM Job for NPM Projects
Create a devstack environment against which a nodejs project may
be tested. Pre and post hooks are delegated directly to npm via
custom commands, which can be configured per job. The Post hook
additionally is passed the path to devstack's generated 'clouds.yaml'
file, which contains credentials for an admin, a non-admin, and
an alternate user.

Example format:

clouds:
  devstack:
    auth: ...
  devstack-admin:
    auth: ...
  devstack-alt:
    auth: ...

Co-Authored-By: Corentin Ardeois <cardeois@internap.com>
Change-Id: I48624786cc5dacf7d94653c0ce269236ba8c442c