Use constraints for venv
Use zuul-cloner with constraints for python jobs that use venv.
Those jobs all run on ubuntu nodes, so zuul-cloner is available.

Note that we can switch using constraints in projects yet since the
translation jobs use venv environment and run on the proposal node that
does not have constraints setup.

Change-Id: I1cd03efc184a7fba6073afd0dd16e6e9f17facd8