Remove the ansible-role 'func' experimental jobs from openstack-ansible
With the implementation of https://review.openstack.org/329892 which
uses the script-builder, instead of tox, the tox-based 'func' jobs
are no longer required.

Unfortunately the tox-based jobs were not suitable for the integrated
build which needs to be executed using the script builder to make it
a reasonable approximation of how a deployment is executed in a
production environment.

Change-Id: I8cf533c1b0e2e5c1c32887da6b567a9b8d90c3f7