puppet: workaround for puppet-ceph
Here is how works puppet-ceph module:
- puppet-ceph stable/hammer works with Puppet OpenStack stable/mitaka
- puppet-ceph master (future stable/jewel) works with Puppet OpenStack
  master (future stable/newton and beyond).

Because of that, we need to checkout puppet-openstack-integration with
the right ZUUL_BRANCH otherwise a patch in puppet-ceph stable/hammer
would try to checkout stable/hammer which doesn't exist and would
fallback to master and CI would fail because stable/hammer works with
Mitaka and not Newton.

This is a workaround for puppet-ceph, we'll get rid of it at the end of
Hammer support.

Change-Id: I257f181d483d63c72b097b6fb25acce352501aa6