Clean up TinyIPA chroot directories in POST job
The TinyIPA post job is currently failing because when it tries to
publish the files, the rsync command attempts to search all directories
for matching source files including the chroot build directories. This
causes it to fail because it doesn't have the permissions to read so of
the directories inside the chroot.

The zuul code comment explaining the behaviour:
https://github.com/openstack-infra/zuul/blob/176431ec1488fbde93c953972f639b8a31d173c8/zuul/launcher/ansiblelaunchserver.py#L939

"include any directories (so that we continue to search for the thing we
want no matter how deep it is),"

Change-Id: I3373d7a6e0220407a84d60b2b2051fa13ffe5174