Re-enable all three jobs for TripleO
Now that rh1 is back up and running we have capacity to run all of
the jobs again.  However, in the interest of not immediately
tripling the load on the cloud, I'm adding the updates job as
experimental for now.  That will allow us to verify it still works
in the OVB environment and make sure we're stable running both the
ha and nonha jobs before we add updates back into the mix.  A
follow-up patch will add it back to the regular check queue.

Change-Id: I8d7ba48b81d91fb6e0d8f61f17ecd40690759d08