Fix Ironic functional job by properly setting PROJECTS
When we construct the environment Ironic functional gate job,
we got the following error:

The /opt/stack/new/ironic-python-agent project was not found;
if this is a gate job, add the project to the $PROJECTS variable
in the job definition.

This patch adds openstack/ironic-python-agent to $PROJECTS
to fix this problem.

Change-Id: I57b886d62c2b6917d2cb0e300d1319d47477e578