Fix incorrect Ironic functional job setting
Change I57b886d62c2b6917d2cb0e300d1319d47477e578 was incorrect.
It added "openstack/ironic-python-agent" in python-ironicclient
functional test job configuration. But it should actually be
added to ironic functional test job configuration.

When we construct the environment Ironic functional gate job,
we got the following error:

The /opt/stack/new/ironic-python-agent project was not found;
if this is a gate job, add the project to the $PROJECTS variable
in the job definition.

This patch adds openstack/ironic-python-agent to $PROJECTS to
fix this problem.

Change-Id: I5f8edbd50b81c897c74beb4021becdaa930d3ea8