fix heat-translator gate failures
Currently, the heat-translator project has a job that is failing [1],
this happened after the job was renamed (the new job now has
"ubuntu-xenial" added to the name. This causes problems with the python
interpreter since it has a limit of 127 characters.

This patch will change the job name to be shorter, it should not impact
the gate since it is currently a non-voting job anyway.

[1] http://logs.openstack.org/26/352126/2/check/gate-heat-translator-tox-py27-tosca-parser-master-ubuntu-xenial/a156262/console.html

Change-Id: I0a44bf05b24a66a76c9dac3f262991c894677ceb