Add devstack test for shade and master of libs
It's good to know about things that client libs do to break us before
they release and break us. Add a devstack job config that installs all
of the client libs from source rather than from pip. This should usually
work, but if it breaks, it means there is something coming that we need
to be prepared for.

Change-Id: Ie387ab3f47d9f79f5d344bcd330784d49198e095