puppet/tripleo: restore ovb stable jobs
gate-tripleo-ci-centos-7-nonha-multinode is a recent job that only works
in master (and future stable/newton), but not on stable/mitaka.

We recently switched some Puppet modules from ovb-nonha to
nonha-multinode but since nonha-multinode is not run in stable branches,
we don't have TripleO jobs anymore for Puppet module stable CI.

This patch adds the tripleo-ovb-non ha jobs for mitaka and liberty on
the zuul layout for puppet modules that have multijob running.

Change-Id: Ie9ef49c3acbfe284ad90b1ff02db53e66535dc92