increase novaclient functional timeout.
During current runs setting up the environment is taking long enough
that the tests are given only 24 minutes to run. This is very often
not long enough for the tests to complete. So we're getting failure
unrelated to code changes.

Move the job timeout from 70 minutes to 120 minutes to give us more
time.

Change-Id: I7060287c77b8a86c656464ee3f9cb5135eb3817e