Set the launchpad name for api-wg
To avoid hogging a rather generic name in launchpad, the API-WG created
a project named 'openstack-api-wg'. There are now bugs being tracked in
that project and the group would like for bug references in commit messages
to update the bugs as possible. Because of the disjunction between gerrit
project name ('api-wg') and launchpad project name this groups addition
is required in project-confg.

There is are openstack-api-wg-bugs and openstack-api-wg-drivers groups on
launchpad.

Change-Id: I6111c63c4365aaa67162e6e278caf0692988eab0