Update trove jobs to include api-ref job
This change adds the api-ref jobs to the trove template. While getting
this to work, another (latent) problem was also discovered. Since the
restrictions on the gate-trove-tox-doc-publish-checkbuild were limited
to changes to the (non existant) apidocs directory, when a previous
change to tox.ini removed the install_command on the checkbuild job,
it was allowed to merge. Added tox.ini to the trigger to re-run that
job.

This change is dependent on the merging of changes in Trove to
actually create this job, etc.,

Change-Id: Ia28c66d38b7a07bd5e0db7a15f49a28377c29330
Depends-On: I3315261aa18729fa7a6aa79d4a1d6c24de1e2c6b