Remove deb-websockify from projects.yaml
The upstream deb-websockify repo is corrupt in a manner that gerrit/jgit
won't accept its objects. Specifically:

  warning in tree 7a80ca084167b04a732640a2f18ec86af55377fb:
  zeroPaddedFilemode: contains zero-padded file modes

This means we can't populate the repo properly which in turn breaks our
image builds when they try to clone and cache this repo. Work around
that for now until the upstream repo can be corrected by removing it
from projects.yaml.

For more information on how to fix repos in this state see:
https://www.perforce.com/blog/150109/git-out-vulnerability-out-fire but
tl:dr is run:

  git fast-export --all | (cd ../myFixedRepo; git fast-import)

Change-Id: I95c9bd5088c5c02440fbf15cd588884b5e0dd954