Remove Neutron Postgres job from the Neutron check queue
We gave this job a go: it was not executed on a continuous basis
for a while; after watching how it perfomed on [1] for at least
a few months we noticed that a) it is rather flaky, b) no-one cares
to bring it back to sanity, and c) our queue may be rather busy.

We should consider having nodes be put to better use and therefore
this patch demotes the job to the experimental queue for occasional
troubleshooting, if deemed necessary.

[1] http://grafana.openstack.org/dashboard/db/neutron-failure-rate?panelId=8&fullscreen

Change-Id: I2cb0effac7cc3c1569555dee31caecaf07b91e86