Bump ubuntu-(trusty|xenial) to 20 min-ready
With over 1.199k nodes available to use. Lets up both trusty and
xenail to 20 again.  This is the first time xenail will be 20, however
ubuntu-trust was decreased in:

  I4215296836b7ec3781cddd80fd2ae224063541d6

so this is a partial revert. Once we fully migrate devstack jobs to
xenial, I believe we could drop trusty down again.

Change-Id: I41876e5850dc96914437dbb068212fdd9eeff550
Signed-off-by: Paul Belanger <pabelanger@redhat.com>