tripleo: align CI scenarios with Puppet CI
TripleO has the desire to cover more OpenStack services testing in the
CI but we also care of saving CI resources.

Puppet CI had the same challenge and we created scenarios, with a
specific set of services deployed by each scenario.

TripleO investigated this way and we successfully reproduce the same
mechanism in our CI.

Now, we would like to align the scenarios from Puppet CI to bring
consistency and re-use the feedback from Puppet community.

This patch is a starter to reproduce Puppet CI:
https://github.com/openstack/puppet-openstack-integration#description

- scenario001 will be a Telemetry scenario with Aodh, Ceilometer and
  Gnocchi with the rest of compute services.
- scenario002 will just deploy Cinder with the rest of compute services.
- scenario003 was scenario001 and will continue to test Sahara with the
  rest of compute services.

More consistency, more coverage, without wasting CI resources.

Change-Id: I9d0c7e12ffbe2afc32a2acad0ce9f0d4e1da4499